Fix 403 error
Problem:
After spinning up the containers, attempting to access the page gives a 403 error.

Cause:
The html folder is empty. This used to work file, but it seems that the recent updates to Docker for Mac have made it so that nested volumes are no longer supported.

Fix:
I changed the html folder from a nested volume into a symbolic link.